Implement delete
Starting to look ugly. Would like to refactor to remove the duplications
in 'insert', 'includes?' and 'delete'. They implement the same (or
similar) logic for how to traverse the tree, and how to do a deep update.

TODO: Also add a 'normalize' function that will re-balance the tree and
remove all 'deleted' nodes.